+++ app/Models/Tag.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Technology extends Model
+class Tag extends Model
 {
     /**
      * The name of the database.
      *
      * @var string
      */
-    protected $table = 'technologies';
+    protected $table = 'tags';
 
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
-    protected $fillable = ['name', 'label'];
+    protected $fillable = ['name'];
 }
